rag-server/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
python-dotenv==1.0.1
langchain==0.1.4
langchain-openai==0.0.5
chromadb==0.4.22
sentence-transformers==2.2.2
python-multipart==0.0.6
tiktoken==0.5.2
numpy==1.26.3
pandas==2.2.0
